Add files via upload
There are many types of computer viruses, trojans, RATs, etc. but one of the most destructive computer viruses I have ever seen is Fork Bomb and the main thing which makes it more terrible is that once the victim device is affected by it then the only option remains is that remove batter of the target device and restart it. It is a simple code that can crash any computer even if it has the latest security patch. This code is known as Fork Bomb. So if you are excited let's get started.